Resolve issue with cached Python not activating
When running Travis tests using a previously cached Python, calls to
`pyenv global $PYENV_VERSION` were not activating the environment. I
think, since in this case the PYENV_VERSION was not actually being
installed with `pyenv install` and rather linked to an existing
directory, the necessary shim wasn't being created. Running `pyenv init`
after creating the link to the version fixed the issue.