Verify build thoroughly before committing to cache
This change attempts to address a problem where a corrupt or otherwise
non-functioning Python build could be saved in the Travis CI cache. If a
Python build fails (e.g., the "pyenv install" command exits with status code
"1"), then the setup-pyenv.sh script exits with status code "1." Also, before
a cached Python is used, it gets verified. If it can't be verified, it's
cleared out of the cache and a new build is installed.