Improve error handling and reporting
The aim of this set of changes is to make the handling and reporting of
errors in setup-pyenv.sh as robust as possible. The script now does the
following. First it attempts to find and use a valid cached Python. If
it can't find the right version in the cache or if the Python build in
the cache is invalid, then it tries reinstalling (deleting a broken
cached existing version first). If the python build fails for any
reason, it immediately returns a status code 1. After a successful
reinstall, it tries moving the ~/.pyenv/versions/ directory to the cache
directory. If the move fails, it issues a warning but continues anyway.
It attempts to validate the newly installed version and returns a status
code 1 if that fails. As before, it does one final validation after the
virtualenv is created.

Messages are now output at each step to make it clearer what's going on
in the Travis logs. If a validation step fails, it outputs some
debugging information to help with troubleshooting.