Configure TravisCI to test building CPython on macOS
These changes remove the global "python:'2.7'" key and adds two new
build configurations for macOS. The key removal is necessary to even get
TravisCI to start a macOS job and 'travis-pyenv' does not need
a TravisCI-provided Python version on the Linux or macOS jobs. The first
configuration tests if the bare-basic installation works for Python
3.6.0 on macOS. The second configuration checks if
a "--enable-framework" (macOS only) build is able to be built. This
second configuration would be useful for PyInstaller builds on macOS.